requirements.txt
# VLESS+Reality VPN Management System - Python Dependencies
# Version: 1.0.0

# Core Telegram Bot Dependencies
python-telegram-bot==20.7
qrcode[pil]==7.4.2
Pillow==10.1.0
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Utility Dependencies
python-dateutil==2.8.2
pytz==2023.3

# Security Dependencies
cryptography==41.0.7
bcrypt==4.1.2
